[iOS]  refactor :recycle: UITableViewCell이 가지고 있던 버튼만드는 메소드를 컨트롤러로 이동
#129
